Still unlikely model but now neurons are dying slowly
Neurons are dying slowly but pro inflammatory astrocytes are not involved. 
It seems more here is a case of build-up of pro inflammatory microglia and dying of both astrocytes and neurons!
Hopefully it is a micro-step in the right direction.
It is also more realistic with a pre-synaptic neuron and an exocytosis between the pre-synaptic and the moto-neuron, as well as dendritic cells and cytokines.